About the typecasting in java

+++ TheTypecasting/Typecasting.java
@@ -1,8 +1,7 @@
 package TheTypecasting;
 
 public class Typecasting {
-    public static void main(String []args)
-    {
+    public static void main(String[] args) {
         /*Typecasting ya implicit typecasting  */
         int a = 45;
         double b;
@@ -18,7 +17,7 @@ public class Typecasting {
         /* Trancation typecasting in java */
         int n1=21;
         int n2=12;
-        float res=n1/n2;
+        double res=n1/n2;
         System.out.println(res);
     }
 }
